requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
uritemplate==4.1.1
